+++ history.md
@@ -5,6 +5,7 @@
 * change function *COL.findObsoleteElements* to handle with Collection now it is possible to use it with List and Set
 * adding function *findObsoleteElementsList*   
 * adding function *findObsoleteElementsSet*
+* adding progressbar option to *StatBar*
 
 #### 2017-11-07 Version 1.0.0
 * first release